Added extra functionality to Point class( an iterator ) and PersistenceDiagram( dimension property and __len__ func ).
persistence-diagram.h:

Added a new read-only dimension member and member function to access it.
With a new constructor that that takes in an int type to initialize dimension.


persistence-diagram.cpp:

Added new bp::init constructor. Takes in an integer type to initialize the dimension. Exposed the dimension property. Exposed the size property via a __len__ method.


__init__.py:

Added an iterator for Point objects. This iterates over the coords and then the data( if present ).